LPS-169073 Move the usage of abstract method doRemoveShardedPortalCache(long, Set<PortalCache<K, V>>) in BasePortalCacheManager to EhcachePortalCacheManager to remove the abstract method from BasePortalCacheManager
